Refactor app handlers into hook
